+++ .github/CODEOWNERS
@@ -1,1 +1,4 @@
 * @iamrecursion @BinarySoftware @joenash
+
+# Repo Configuration
+/.github/settings.yml @iamrecursion
