+++ .github/workflows/python-app.yml
@@ -20,16 +20,9 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: 3.9
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install flake8 PyExecJS PyGithub
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Set up Poetry
+      uses: Gr1N/setup-poetry@v4
     - name: Run app
       run: |
-        python main.py --token=${{ secrets.GITHUB_TOKEN }}
+        poetry install
+        poetry run python src/main.py --token=${{ secrets.GITHUB_TOKEN }}
